requirements.txt
asgiref==3.4.1
brython==3.8.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
coreapi==2.3.3
coreschema==0.0.4
crypto==1.4.1
cryptography==35.0.0
defusedxml==0.7.1
Django==3.2.8
django-cart==1.0.4
django-cors-headers==3.10.1
django-filter==21.1
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.20.0
idna==3.3
importlib-metadata==1.7.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
Markdown==3.3.4
MarkupSafe==2.0.1
mysqlclient==2.0.3
Naked==0.1.31
oauthlib==3.2.0
openapi-codec==1.3.2
packaging==21.3
Pillow==8.4.0
psycopg2-binary==2.9.3
pycparser==2.20
PyJWT==2.3.0
PyMySQL==1.0.2
pyparsing==3.0.6
python3-openid==3.2.0
pytz==2021.3
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.1
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
shellescape==3.8.1
simplejson==3.17.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
stripe==2.67.0
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.7
vuepy==0.3.0
zipp==3.6.0
